LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-construction.any.worker-expected.txt


FAIL Test we can construct a VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test closed VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test we can construct a VideoFrame with a negative timestamp. Can't find variable: OffscreenCanvas
FAIL Test that timestamp is required when constructing VideoFrame from ImageBitmap promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
FAIL Test that timestamp is required when constructing VideoFrame from OffscreenCanvas promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: OffscreenCanvas"
PASS Test that timestamp is NOT required when constructing VideoFrame from another VideoFrame
FAIL Test we can construct an odd-sized VideoFrame. Can't find variable: OffscreenCanvas
PASS Test constructing w/ unusable image argument throws: HAVE_NOTHING <video>.
FAIL Test constructing w/ unusable image argument throws: emtpy Canvas. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: closed ImageBitmap. Can't find variable: OffscreenCanvas
FAIL Test constructing w/ unusable image argument throws: closed VideoFrame. Can't find variable: OffscreenCanvas
FAIL Test invalid CanvasImageSource constructed VideoFrames Can't find variable: OffscreenCanvas
PASS Test visibleRect metadata override where source display size = visible size
FAIL Test visibleRect metadata override where source display width = 2 * visible width (anamorphic) assert_equals: cropRightFrame.displayWidth expected 4 but got 2
FAIL Test visibleRect metadata override where source display size = 2 * visible size for both width and height assert_equals: cropRightFrame.displayWidth expected 4 but got 2
FAIL Test visibleRect + display size metadata override Can't find variable: OffscreenCanvas
FAIL Test display size metadata override Can't find variable: OffscreenCanvas
FAIL Test invalid buffer constructed VideoFrames assert_throws_js: odd coded height function "() => constructFrame({timestamp: 1234, codedWidth: 4, codedHeight: 1})" did not throw
PASS Test Uint8Array(ArrayBuffer) constructed I420 VideoFrame
PASS Test ArrayBuffer constructed I420 VideoFrame
PASS Test planar constructed I420 VideoFrame with colorSpace
FAIL Test buffer constructed I420+Alpha VideoFrame VideoPixelFormat is not supported
PASS Test buffer constructed NV12 VideoFrame
FAIL Test buffer constructed RGB VideoFrames VideoPixelFormat is not supported
FAIL Test VideoFrame constructed VideoFrame Can't find variable: OffscreenCanvas
FAIL Test we can construct a VideoFrame from an offscreen canvas. Can't find variable: OffscreenCanvas
PASS Test I420 VideoFrame with odd visible size
FAIL Test I420A VideoFrame and alpha={keep,discard} VideoPixelFormat is not supported
FAIL Test RGBA, BGRA VideoFrames with alpha={keep,discard} assert_equals: plane format expected "RGBX" but got "RGBA"
FAIL Test a VideoFrame constructed from canvas can drop the alpha channel. Can't find variable: OffscreenCanvas
